refactor(api-reference): extract overview helpers, types, and styles into dedicated modules
Move API reference overview utilities, types, callout-block parsing, and
sx style objects out of `[slug].tsx` (1394 → 826 lines) and into modules
that follow the repo's existing conventions (`src/utils/`, `src/styles/`).

- Add generic string helpers to `src/utils/string-utils.ts`
  (`normalizeWhitespace`, `stripHTML`, `getFirstSentence`, `trimToLength`,
  `removeTitlePrefix`).
- Add `src/utils/replaceCalloutBlocks.ts` for `[block:callout]`
  Markdown/HTML transforms (kept separate from `replaceMagicBlocks.ts`
  because that one emits inline JSX for guide pages).
- Add `src/utils/api-reference-overview.ts` for feature-specific helpers
  and types (`OverviewEndpoint*`, `getOverviewEndpointHash`,
  `buildOverviewMetaDescription`, `buildOverviewEndpointGroups`,
  `getEndpointPathFromLocation`).
- Add `src/styles/api-reference.ts` for the page sx style objects and
  `getOverviewEndpointMethodBadgeSx`, mirroring `src/styles/openapi-preview.ts`.

No behavior change; references in `[slug].tsx` updated to use the new
modules.
